third_party/WebKit/LayoutTests/http/tests/devtools/coverage/coverage-view-filter.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/elements-test.js"></script>
<script src="../resources/coverage-test.js"></script>
<script src="resources/coverage.js"></script>

<link rel="stylesheet" type="text/css" href="resources/highlight-in-source.css">

<script>
async function test() {
  CoverageTestRunner.startCoverage();
  await TestRunner.evaluateInPagePromise('performActions()');
  await CoverageTestRunner.stopCoverage();

  var coverageView = self.runtime.sharedInstance(Coverage.CoverageView);
  setFilter('devtools');
  CoverageTestRunner.dumpCoverageListView();
  setFilter('CES/COV');
  CoverageTestRunner.dumpCoverageListView();
  setFilter('no pasaran');
  CoverageTestRunner.dumpCoverageListView();
  setFilter('');
  CoverageTestRunner.dumpCoverageListView();
  TestRunner.completeTest();

  function setFilter(text) {
    coverageView._filterInput.setValue(text);
    coverageView._filterInput._onChangeCallback();
    TestRunner.addResult(`Filter: ${text}`);
  }
}

</script>
</head>

<body onload="runTest()">
<p class="class">
Tests the filter is properly applied to coverage list view.
</p>

</body>
</html>
