third_party/WebKit/LayoutTests/http/tests/devtools/coverage/coverage-view-filter-expected.txt

Tests the filter is properly applied to coverage list view.

Filter: devtools
http://127.0.0.1:8000/devtools/coverage/resources/coverage.js JS used: 235 unused: 167 total: 402
http://127.0.0.1:8000/devtools/coverage/resources/highlight-in-source.css CSS used: 67 unused: 142 total: 209
Filter: CES/COV
http://127.0.0.1:8000/devtools/coverage/resources/coverage.js JS used: 235 unused: 167 total: 402
Filter: no pasaran
Filter: 
http://127.0.0.1:8000/devtools/coverage/resources/coverage.js JS used: 235 unused: 167 total: 402
http://127.0.0.1:8000/devtools/coverage/resources/highlight-in-source.css CSS used: 67 unused: 142 total: 209
